# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alliance University Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FEES ────────────────────────────────────────────────────────────────────
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Create fee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/{id}:
    get:
      operationId: getFeeRecord
      tags: [fees]
      summary: Get fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fee record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
    patch:
      operationId: updateFeeRecord
      tags: [fees]
      summary: Update fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
...
  /fees/{id}/receipt:
    get:
      operationId: getFeeReceipt
      tags: [fees]
      summary: Get fee receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeReceipt"

  # ─── ATTENDANCE ──────────────────────────────────────────────────────────────
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List results
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      summary: Create result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get result details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: targetRole
          in: query
          schema:
            type: string
            enum: [all, student, staff, admin]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    patch:
      operationId: updateNotification
      tags: [notifications]
      summary: Update notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DOCUMENTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StaffMember:
      type: object
      required: [id, name, employeeId, department, designation, email, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        employeeId:
          type: string
        department:
          type: string
        designation:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, employeeId, department, designation, email, phone]
      properties:
        name:
          type: string
        employeeId:
          type: string
        department:
          type: string
        designation:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        password:
          type: string
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        designation:
          type: string
        email:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
    FeeRecord:
      type: object
      required: [id, studentId, semester, totalFees, paidAmount, pendingAmount, status, dueDate]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        semester:
          type: integer
        academicYear:
          type: ["string", "null"]
        totalFees:
          type: number
        paidAmount:
          type: number
        pendingAmount:
          type: number
        status:
          type: string
          enum: [paid, partial, pending]
        dueDate:
          type: string
        paidDate:
          type: ["string", "null"]
        paymentMode:
          type: ["string", "null"]
        receiptNo:
          type: ["string", "null"]
        createdAt:
          type: string
    FeeInput:
      type: object
      required: [studentId, semester, totalFees, paidAmount, dueDate]
      properties:
        studentId:
          type: integer
        semester:
          type: integer
        academicYear:
          type: string
        totalFees:
          type: number
        paidAmount:
          type: number
        dueDate:
          type: string
        paymentMode:
          type: string
        receiptNo:
          type: string
    FeeUpdate:
      type: object
      properties:
        paidAmount:
          type: number
        status:
          type: string
          enum: [paid, partial, pending]
        paidDate:
          type: string
        paymentMode:
          type: string
        receiptNo:
          type: string
...
    FeeReceipt:
      type: object
      required: [id, receiptNo, receiptDate, studentName, fatherName, enrollmentNo, admissionNo, course, totalAmount, particulars]
      properties:
        id:
          type: integer
        receiptNo:
          type: string
        receiptDate:
          type: string
        studentName:
          type: string
        fatherName:
          type: string
        enrollmentNo:
          type: string
        admissionNo:
          type: string
        course:
          type: string
        address:
          type: ["string", "null"]
        totalAmount:
          type: number
        amountInWords:
          type: string
        paymentMode:
          type: string
        bankName:
          type: ["string", "null"]
        particulars:
          type: array
          items:
            $ref: "#/components/schemas/FeeParticular"
    FeeParticular:
      type: object
      required: [name, amount]
      properties:
        name:
          type: string
        amount:
          type: number
...
    Result:
      type: object
      required: [id, studentId, semester, academicYear, sgpa, cgpa, status, subjects, totalCredits]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        semester:
          type: integer
        academicYear:
          type: string
        sgpa:
          type: number
        cgpa:
          type: number
        status:
          type: string
          enum: [pass, fail, withheld]
        totalCredits:
          type: integer
        equivalentPercentage:
          type: ["number", "null"]
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectResult"
        resultDate:
          type: ["string", "null"]
    ResultInput:
      type: object
      required: [studentId, semester, academicYear, sgpa, cgpa, status, subjects, totalCredits]
      properties:
        studentId:
          type: integer
        semester:
          type: integer
        academicYear:
          type: string
        sgpa:
          type: number
        cgpa:
          type: number
        status:
          type: string
          enum: [pass, fail, withheld]
        totalCredits:
          type: integer
        equivalentPercentage:
          type: number
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectResult"
        resultDate:
          type: string
    SubjectResult:
      type: object
      required: [subjectCode, subjectName, credits, iaMarks, eaMarks, totalMarks, maxIa, maxEa, maxTotal, grade, gradePoint]
      properties:
        subjectCode:
          type: string
        subjectName:
          type: string
        credits:
          type: integer
        iaMarks:
          type: integer
        eaMarks:
          type: integer
        totalMarks:
          type: integer
        maxIa:
          type: integer
        maxEa:
          type: integer
        maxTotal:
          type: integer
        grade:
          type: string
        gradePoint:
          type: integer
    Notification:
      type: object
      required: [id, title, message, targetRole, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        targetRole:
          type: string
          enum: [all, student, staff, admin]
        priority:
          type: string
          enum: [normal, important, urgent]
        createdAt:
          type: string
        createdBy:
          type: ["string", "null"]
    NotificationInput:
      type: object
      required: [title, message, targetRole]
      properties:
        title:
          type: string
        message:
          type: string
        targetRole:
          type: string
          enum: [all, student, staff, admin]
        priority:
          type: string
          enum: [normal, important, urgent]
    NotificationUpdate:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        targetRole:
          type: string
          enum: [all, student, staff, admin]
        priority:
          type: string
          enum: [normal, important, urgent]
...