# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alliance University Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  name: string;
  employeeId: string;
  department: string;
  designation: string;
  email: string;
  phone: string;
  photoUrl?: string;
  password?: string;
}